backend/requirements.txt
Flask
flask-cors
python-dotenv
pymongo
openai
PyMuPDF
